fix: Change lines cleared type from Int to Long
Changes the data type for the number of lines cleared from `Int` to `Long` to prevent potential overflow issues.

This update is applied consistently across the following areas:
- `GameComponent.State.finalLinesCleared`
- `DefaultGameComponent.DialogConfig.GameOver.lines`
- The `lines` parameter in the `GameStatsRow` composable function.

+++ composeApp/src/commonMain/kotlin/com/yet/tetris/ui/screens/game/GameScreen.kt
@@ -184,7 +184,7 @@ private fun GamePlayingContent(
 
 private fun GameStatsRow(
     score: Long,
-    lines: Int,
+    lines: Long,
     time: Long
 ) {
     Card(
